Introduce scorep.instrumenter.utils module
